DAOS-12638 tse: increase lock granularity to all progress call
The tse lock should be applied overall during the progress call and when
the body function, prep, and completion cbs are being executed.
In a multi-threaded application, without this lock, the app ends up
crashing with memory heap corruption in the client stack which manifests
in different assertions and segfaults.

Required-githooks: true

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>